Move HyperLogLogPlusPlus to enterprise package
