+++ enterprise/functions/src/main/java/io/crate/execution/engine/aggregation/impl/HyperLogLogPlusPlus.java
@@ -1,23 +1,24 @@
 /*
- * Licensed to Elasticsearch under one or more contributor
- * license agreements. See the NOTICE file distributed with
- * this work for additional information regarding copyright
- * ownership. Elasticsearch licenses this file to you under
- * the Apache License, Version 2.0 (the "License"); you may
- * not use this file except in compliance with the License.
- * You may obtain a copy of the License at
+ * This file is part of a module with proprietary Enterprise Features.
  *
- *    http://www.apache.org/licenses/LICENSE-2.0
+ * Licensed to Crate.io Inc. ("Crate.io") under one or more contributor
+ * license agreements.  See the NOTICE file distributed with this work for
+ * additional information regarding copyright ownership.
  *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
+ * Unauthorized copying of this file, via any medium is strictly prohibited.
+ *
+ * To use this file, Crate.io must have given you permission to enable and
+ * use such Enterprise Features and you must have a valid Enterprise or
+ * Subscription Agreement with Crate.io.  If you enable or use the Enterprise
+ * Features, you represent and warrant that you have a valid Enterprise or
+ * Subscription Agreement with Crate.io.  Your use of the Enterprise Features
+ * if governed by the terms and conditions of your Enterprise or Subscription
+ * Agreement with Crate.io.
+ *
+ * This is based on the ASL 2.0 licensed HyperLogLogPlus implementation found in Elasticsearch.
  */
 
-package org.elasticsearch.search.aggregations.metrics.cardinality;
+package io.crate.execution.engine.aggregation.impl;
 
 import org.apache.lucene.util.packed.PackedInts;
 import org.elasticsearch.common.io.stream.StreamInput;
@@ -134,7 +135,8 @@ public final class HyperLogLogPlusPlus {
 
     private boolean algorithm = LINEAR_COUNTING;
     private final Hashset hashSet;
-    private final int p, m;
+    private final int p;
+    private final int m;
     private final double alphaMM;
     private final byte[] runLens;
 
@@ -151,15 +153,15 @@ public final class HyperLogLogPlusPlus {
         hashSet = new Hashset();
         final double alpha;
         switch (p) {
-        case 4:
-            alpha = 0.673;
-            break;
-        case 5:
-            alpha = 0.697;
-            break;
-        default:
-            alpha = 0.7213 / (1 + 1.079 / m);
-            break;
+            case 4:
+                alpha = 0.673;
+                break;
+            case 5:
+                alpha = 0.697;
+                break;
+            default:
+                alpha = 0.7213 / (1 + 1.079 / m);
+                break;
         }
         alphaMM = alpha * m * m;
     }
